requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.3.0
python-multipart
jinja2
aiofiles
psycopg2-binary
python-dotenv
